alter data type for rupiah storage

+++ database/migrations/2024_04_22_071843_create_purchases_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('purchases', function (Blueprint $table) {
             $table->id();
             $table->string('nama_barang');
-            $table->bigInteger('qty');
-            $table->decimal('harga_beli');
+            $table->integer('qty');
+            $table->bigInteger('harga_beli');
             $table->timestamps();
         });
     }
